Update provider and model list Move bing.create_images and cookies helper Disable some providers
