+++ g4f/Provider/GptChatly.py
@@ -10,7 +10,6 @@ class GptChatly(AsyncProvider):
     working = True
     supports_message_history = True
     supports_gpt_35_turbo = True
-    supports_gpt_4 = True
 
     @classmethod
     async def create_async(
