+++ g4f/Provider/__init__.py
@@ -53,6 +53,8 @@ from .Vercel          import Vercel
 from .Ylokh           import Ylokh
 from .You             import You
 
+from .CreateImagesBing import CreateImagesBing
+
 import sys
 
 __modules__: list = [
