+++ g4f/Provider/bing/create_images.py
@@ -7,10 +7,9 @@ from __future__ import annotations
 import asyncio
 import time
 import json
-import os
 from aiohttp import ClientSession, BaseConnector
 from urllib.parse import quote
-from typing import Generator, List, Dict
+from typing import List, Dict
 
 try:
     from bs4 import BeautifulSoup
@@ -19,14 +18,11 @@ except ImportError:
     has_requirements = False
 
 from ..create_images import CreateImagesProvider
-from ..helper import get_cookies, get_connector
-from ...webdriver import WebDriver, get_driver_cookies, get_browser
+from ..helper import get_connector
 from ...base_provider import ProviderType
-from ...image import ImageResponse
-from ...errors import MissingRequirementsError, MissingAuthError
+from ...errors import MissingRequirementsError
 
 BING_URL = "https://www.bing.com"
-TIMEOUT_LOGIN = 1200
 TIMEOUT_IMAGE_CREATION = 300
 ERRORS = [
     "this prompt is being reviewed",
@@ -38,24 +34,6 @@ BAD_IMAGES = [
     "https://r.bing.com/rp/in-2zU3AJUdkgFe7ZKv19yPBHVs.png",
     "https://r.bing.com/rp/TX9QuO3WzcCJz1uaaSwQAz39Kb0.jpg",
 ]
-
-def wait_for_login(driver: WebDriver, timeout: int = TIMEOUT_LOGIN) -> None:
-    """
-    Waits for the user to log in within a given timeout period.
-
-    Args:
-        driver (WebDriver): Webdriver for browser automation.
-        timeout (int): Maximum waiting time in seconds.
-
-    Raises:
-        RuntimeError: If the login process exceeds the timeout.
-    """
-    driver.get(f"{BING_URL}/")
-    start_time = time.time()
-    while not driver.get_cookie("_U"):
-        if time.time() - start_time > timeout:
-            raise RuntimeError("Timeout error")
-        time.sleep(0.5)
 
 def create_session(cookies: Dict[str, str], proxy: str = None, connector: BaseConnector = None) -> ClientSession:
     """
@@ -170,67 +148,6 @@ def read_images(html_content: str) -> List[str]:
         raise RuntimeError("No images found")
     return images
 
-def get_cookies_from_browser(proxy: str = None) -> dict[str, str]:
-    """
-    Retrieves cookies from the browser using webdriver.
-
-    Args:
-        proxy (str, optional): Proxy configuration.
-
-    Returns:
-        dict[str, str]: Retrieved cookies.
-    """
-    with get_browser(proxy=proxy) as driver:
-        wait_for_login(driver)
-        time.sleep(1)
-        return get_driver_cookies(driver)
-
-class CreateImagesBing:
-    """A class for creating images using Bing."""
-
-    def __init__(self, cookies: dict[str, str] = {}, proxy: str = None) -> None:
-        self.cookies = cookies
-        self.proxy = proxy
-
-    def create_completion(self, prompt: str) -> Generator[ImageResponse, None, None]:
-        """
-        Generator for creating imagecompletion based on a prompt.
-
-        Args:
-            prompt (str): Prompt to generate images.
-
-        Yields:
-            Generator[str, None, None]: The final output as markdown formatted string with images.
-        """
-        cookies = self.cookies or get_cookies(".bing.com", False)
-        if "_U" not in cookies:
-            login_url = os.environ.get("G4F_LOGIN_URL")
-            if login_url:
-                yield f"Please login: [Bing]({login_url})\n\n"
-            try:
-                self.cookies = get_cookies_from_browser(self.proxy)
-            except MissingRequirementsError as e:
-                raise MissingAuthError(f'Missing "_U" cookie. {e}')
-        yield asyncio.run(self.create_async(prompt))
-
-    async def create_async(self, prompt: str) -> ImageResponse:
-        """
-        Asynchronously creates a markdown formatted string with images based on the prompt.
-
-        Args:
-            prompt (str): Prompt to generate images.
-
-        Returns:
-            str: Markdown formatted string with images.
-        """
-        cookies = self.cookies or get_cookies(".bing.com", False)
-        if "_U" not in cookies:
-            raise MissingAuthError('Missing "_U" cookie')
-        proxy = os.environ.get("G4F_PROXY")
-        async with create_session(cookies, proxy) as session:
-            images = await create_images(session, prompt, self.proxy)
-            return ImageResponse(images, prompt, {"preview": "{image}?w=200&h=200"})
-
 def patch_provider(provider: ProviderType) -> CreateImagesProvider:
     """
     Patches a provider to include image creation capabilities.
@@ -241,6 +158,7 @@ def patch_provider(provider: ProviderType) -> CreateImagesProvider:
     Returns:
         CreateImagesProvider: The patched provider with image creation capabilities.
     """
+    from ..CreateImagesBing import CreateImagesBing
     service = CreateImagesBing()
     return CreateImagesProvider(
         provider,
