+++ g4f/models.py
@@ -15,15 +15,12 @@ from .Provider   import (
     DeepInfra,
     ChatBase,
     Liaobots,
-    GeekGpt,
-    FakeGpt,
     FreeGpt,
     Llama2,
-    Vercel, 
-    Phind,
+    Vercel,
+    Gemini,
     GptGo,
     Gpt6,
-    Bard, 
     Bing,
     You,
     Pi,
@@ -53,7 +50,7 @@ default = Model(
     base_provider = "",
     best_provider = RetryProvider([
         Bing,
-        ChatgptAi, GptGo, GeekGpt,
+        ChatgptAi, GptGo,
         You,
         Chatgpt4Online
     ])
@@ -65,7 +62,6 @@ gpt_35_long = Model(
     base_provider = 'openai',
     best_provider = RetryProvider([
         FreeGpt, You,
-        GeekGpt, FakeGpt,
         Chatgpt4Online,
         ChatgptDemoAi,
         ChatgptNext,
@@ -168,10 +164,10 @@ openchat_35 = Model(
 )
 
 # Bard
-bard = palm = Model(
-    name          = 'palm',
+gemini = bard = palm = Model(
+    name          = 'gemini',
     base_provider = 'google',
-    best_provider = Bard
+    best_provider = Gemini
 )
 
 claude_v2 = Model(
@@ -264,8 +260,8 @@ class ModelUtils:
         'airoboros-70b': airoboros_70b,
         'airoboros-l2-70b': airoboros_l2_70b,
         'openchat_3.5': openchat_35,
+        'gemini': gemini,
         'gemini-pro': gemini_pro,
-        'bard': bard,
         'claude-v2': claude_v2,
         'pi': pi
     }
